+++ src/openvpn/mtu.h
@@ -80,6 +80,11 @@
 #define MSSFIX_DEFAULT     1492
 
 /*
+ * Default maximum size of control channel packets
+ */
+#define TLS_MTU_DEFAULT    1250
+
+/*
  * Alignment of payload data such as IP packet or
  * ethernet frame.
  */